[docs] Move notifications to own component
